agw: fix outbound AGWPE port numbering for K/U frames
AGWPE ports are zero-based while graywolf channel IDs are one-based, so
BroadcastMonitoredUI and BroadcastRawKISS were both announcing frames
one port higher than configured (e.g. channel 1 as port 1 instead of
port 0). BroadcastRawKISS also hardcoded its leading TNC-indicator byte
to 0 instead of deriving it from the port.

Add Server.portFor, the inverse of cfg.PortToChannel, and use it for
both the header Port field and the KISS leading byte so they agree
with each other and with the rest of graywolf's port/channel mapping.

Fixes #588.

+++ pkg/app/rxfanout.go
@@ -168,12 +168,12 @@ func (a *App) dispatchRxFrame(ctx context.Context, item rxFanoutItem, aprsSubmit
 	}
 
 	if srv := a.currentAgwServer(); srv != nil {
-		srv.BroadcastRawKISS(uint8(rf.Channel), rf.Data)
+		srv.BroadcastRawKISS(rf.Channel, rf.Data)
 	}
 
 	if f.IsUI() {
 		if srv := a.currentAgwServer(); srv != nil {
-			srv.BroadcastMonitoredUI(uint8(rf.Channel), f)
+			srv.BroadcastMonitoredUI(rf.Channel, f)
 		}
 		a.digi.Handle(ctx, rf.Channel, f, src)
 		if pkt, err := aprs.Parse(f); err == nil && pkt != nil {
